Make the website dynamical
