+++ article.html
@@ -26,15 +26,16 @@
     <main>
       <section id="article-section">
         <article class="card--article">
-          <div class="card">
+          <div class="card--article-wrapper">
             <div class="card--article__img">
-              <img
+              <!-- <img
                 class="img-fluid"
                 src="https://cdn.prod.www.spiegel.de/images/f07fe829-de28-40ab-8db1-91f2e3373ebf_w948_r1.77_fpx28.63_fpy44.98.jpg"
                 alt="Bett"
-              />
+              /> -->
             </div>
-            <h2>Wie Frankreich seine Bettwanzen-Plage bekämpft</h2>
+            <!-- <h2>Wie Frankreich seine Bettwanzen-Plage bekämpft</h2> -->
+
             <p>
               Sie beißen Bürger und Touristen, lösen Juckreiz und großes
               Unbehagen aus: Frankreich hat ein Problem mit Bettwanzen. Im Kampf
@@ -109,5 +110,6 @@
           <i class="fas fa-user-cog bottom-nav__icon"></i></button
       ></a>
     </footer>
+    <script type="module" src="src/article.js"></script>
   </body>
 </html>
