+++ src/article.js
@@ -1,0 +1,8 @@
+import { articlePicture } from "./data/elements.js";
+
+const picture = document.createElement("img");
+const pictureWrapper = document.querySelector(".card--article__img");
+
+picture.src = articlePicture.imgLink;
+
+pictureWrapper.appendChild(picture);
