+++ src/data/elements.js
@@ -1,0 +1,40 @@
+export const teasers = [
+  {
+    title: "Wie Frankreich seine Bettwanzen-Plage bekämpft",
+    subtitle: "Panorama",
+    imgLink:
+      "https://cdn.prod.www.spiegel.de/images/f07fe829-de28-40ab-8db1-91f2e3373ebf_w948_r1.77_fpx28.63_fpy44.98.jpg",
+    teaserText:
+      "Sie beißen Bürger und Touristen, lösen Juckreiz und großes Unbehagen aus: Frankreich hat ein Problem mit Bettwanzen.Im Kampf gegen das Ungeziefer schaltet sich nun der Staat ein.",
+    teaserLink: "article.html"
+  },
+  {
+    title: "Erst der Mensch, dann der Sport, dann das Drumherum",
+    subtitle: "Malaika Mihambo",
+    imgLink: "https://www.sueddeutsche.de/image/sz.1.4814389/600x337",
+    teaserText:
+      "Weitsprung-Weltmeisterin Malaika Mihambo scheint zu Beginn der neuen Leichtathletik- Saison fast alles zu gelingen.Auch, weil sie klare Prioritäten setzt.",
+    teaserLink: "article-2.html"
+  },
+  {
+    title: "Gehalt: Wie viel verdient eine Chemikantin?",
+    subtitle: "4916 Euro brutto für die Chemikantin",
+    imgLink: "https://media-cdn.sueddeutsche.de/image/sz.1.4815786/215x120",
+    teaserText:
+      "Melissa, 23, überwacht Anlagen beim Chemiekonzern BASF und arbeitet in Zwölf-Stunden-Schichten.",
+    teaserLink: "article-3.html"
+  },
+  {
+    title: "Das beste aus unserer Zeitung",
+    subtitle: "Mit Schnitzel, bitte",
+    imgLink: "https://www.sueddeutsche.de/image/sz.1.4813883/208x156",
+    teaserText:
+      "Die Deutschen lieben Krapfen. Die Bäcker wissen das und werden von Jahr zu Jahr einfallsreicher.",
+    teaserLink: "article-4.html"
+  }
+];
+
+export const articlePicture = {
+  imgLink:
+    "https://cdn.prod.www.spiegel.de/images/f07fe829-de28-40ab-8db1-91f2e3373ebf_w948_r1.77_fpx28.63_fpy44.98.jpg"
+};
