--- src/data/teaser.js
@@ -1,20 +1,0 @@
-export const teasers = [
-  {
-    title: "Wie Frankreich seine Bettwanzen-Plage bekämpft",
-    subtitle: "Panorama",
-    imgLink:
-      "https://cdn.prod.www.spiegel.de/images/f07fe829-de28-40ab-8db1-91f2e3373ebf_w948_r1.77_fpx28.63_fpy44.98.jpg",
-    teaserText:
-      "Sie beißen Bürger und Touristen, lösen Juckreiz und großes Unbehagen aus: Frankreich hat ein Problem mit Bettwanzen.Im Kampf gegen das Ungeziefer schaltet sich nun der Staat ein.",
-    teaserLink: "article.html"
-  },
-  {
-    title: "Erst der Mensch, dann der Sport, dann das Drumherum",
-    subtitle: "Malaika Mihambo",
-    imgLink:
-      "https://cdn.prod.www.spiegel.de/images/f07fe829-de28-40ab-8db1-91f2e3373ebf_w948_r1.77_fpx28.63_fpy44.98.jpg",
-    teaserText:
-      "Weitsprung-Weltmeisterin Malaika Mihambo scheint zu Beginn der neuen Leichtathletik- Saison fast alles zu gelingen.Auch, weil sie klare Prioritäten setzt.",
-    teaserLink: "article-2.html"
-  }
-];
